feat: #9 カスタムイベントを定義してmixinから処理を外出し
