+++ src/main/java/me/enchan/minerstools/MinersToolsMod.java
@@ -8,6 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import me.enchan.minerstools.bulk_break.BulkBreakDispatcher;
+import me.enchan.minerstools.events.MinersToolsMainHandToolBreakEvent;
 import me.enchan.minerstools.payloads.MinersToolsModeTogglePayload;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.player.PlayerBlockBreakEvents;
@@ -41,6 +42,46 @@ public class MinersToolsMod implements ModInitializer {
             BulkBreakDispatcher.onBreakBlock((ServerWorld) world, origin, player, state);
         });
 
+        MinersToolsMainHandToolBreakEvent.EVENT.register(player -> {
+            if (!toolStatusByPlayer.getOrDefault(player.getUuid(), false)) {
+                return;
+            }
+
+            // 壊れたアイテムを取得
+            var brokenItemStack = player.getMainHandStack();
+
+            var inventory = player.getInventory().main;
+            var candidateIndex = -1;
+            for (int i = 0; i < inventory.size(); i++) {
+                var itemStack = inventory.get(i);
+                if (itemStack.equals(brokenItemStack)) {
+                    continue;
+                }
+
+                if (!itemStack.getItem().equals(brokenItemStack.getItem())) {
+                    continue;
+                }
+
+                // TODO: より耐久の残っているものから利用していくなどの最適化もできるかも
+                candidateIndex = i;
+            }
+
+            if (candidateIndex < 0) {
+                player.sendMessage(Text.literal("No alternative tools!"), true);
+                return;
+            }
+            var alternativeItem = inventory.get(candidateIndex);
+
+            // 持ち替え
+            var currentSlot = player.getInventory().selectedSlot;
+            inventory.set(candidateIndex, inventory.get(currentSlot));
+            inventory.set(currentSlot, alternativeItem);
+
+            player.currentScreenHandler.sendContentUpdates();
+
+            player.sendMessage(Text.literal("Switched to new one!"), true);
+        });
+
         ServerPlayNetworking.registerGlobalReceiver(MinersToolsModeTogglePayload.ID, (payload, context) -> {
             ServerPlayerEntity player = context.player();
 
@@ -55,5 +96,6 @@ public class MinersToolsMod implements ModInitializer {
                             : Text.literal("OFF").formatted(Formatting.RED));
             player.sendMessage(message, true);
         });
+
     }
 }
